move random to separate file
