+++ neurogym/utils/tasktools.py
@@ -34,72 +34,6 @@ def get_periods_idx(dt, periods):
     return t, {k: get_idx(t, v) for k, v in periods.items() if k != 'tmax'}
 
 
-def uniform(rng, dt, xmin, xmax):
-    if xmin >= xmax:  # the > is to avoid issues when making xmin as big as dt
-        return (xmax//dt)*dt
-    else:
-        return (rng.uniform(xmin, xmax)//dt)*dt
-
-
-def trunc_exp(rng, dt, mean, xmin=0, xmax=np.inf):
-    """
-    function for generating period durations that are multiples of the timestep
-    """
-    if xmin >= xmax:  # the > is to avoid issues when making xmin as big as dt
-        return (xmax//dt)*dt
-    else:
-        while True:
-            x = rng.expovariate(1/mean)
-            if xmin <= x < xmax:
-                return (x//dt)*dt
-
-
-def trunc_exp_new(rng, mean, xmin=0, xmax=np.inf):
-    """
-    function for generating period durations
-    """
-    if xmin >= xmax:  # the > is to avoid issues when making xmin as big as dt
-        return xmax
-    else:
-        while True:
-            x = rng.exponential(mean)
-            if xmin <= x < xmax:
-                return x
-
-
-def random_number_fn(dist, args, rng):
-    """Return a random number generating function from a distribution."""
-    if dist == 'uniform':
-        return lambda: rng.uniform(*args)
-    elif dist == 'choice':
-        return lambda: rng.choice(args)
-    elif dist == 'truncated_exponential':
-        return lambda: trunc_exp_new(rng, *args)
-    elif dist == 'constant':
-        return lambda: args
-    else:
-        raise ValueError('Unknown dist:', str(dist))
-
-
-def random_number_name(dist, args):
-    """Return a string explaining the dist and args."""
-    if dist == 'uniform':
-        return dist + ' between ' + str(args[0]) + ' and ' + str(args[1])
-    elif dist == 'choice':
-        return dist + ' within ' + str(args)
-    elif dist == 'truncated_exponential':
-        string = 'truncated exponential with mean ' + str(args[0])
-        if len(args) > 1:
-            string += ', min ' + str(args[1])
-        if len(args) > 2:
-            string += ', max ' + str(args[2])
-        return string
-    elif dist == 'constant':
-        return dist + ' ' + str(args)
-    else:
-        raise ValueError('Unknown dist:', str(dist))
-
-
 def minmax_number(dist, args):
     """Given input to the random_number_fn function, return min and max."""
     if dist == 'uniform':
